refactor: Remove unneeded wmts-playground
